Revert submodules
Submodules cannot have their own buildSrc, which is needed for the code gen.
We could use multiple packages instead and have buildSrc as an upstream package but that would negate the effect of having the code for different modules live in separate packges.
Going back to single module will make things a lot easier.